crash: Collection Status Enum: initial setup
Introduces CrashCollectionStatus enum, replacing the old
CrashCollector::ErrorType.

CrashCollectionStatus is a more detailed enum -- the combination of
CrashCollectionStatus and the collector should give a good idea of what
code path we followed when collecting a crash (or failing to collect a
crash). Eventually, CrashCollectionStatus will be propagated all the way
up to the main function. There, it will be used in the structured
metrics / CrOSEvents "CrashReporting.CrashReporterStatus" event's Status
field to indicate if we successfully collected a crash, and if not,
why not.

But for this change, I am just introducing the new enum and replacing
ErrorType, and updating other functions in CrashCollector (especially
FinishCrash) to use it. Future CLs will update the collection functions
to propagate those return values to the main function.

Other than some minor changes to error strings and the like, this CL
should have NO impact on the program's behavior.

BUG=b:177552411
TEST=sleep 100 &; kill -SEGV %1; ensured resulting crash looked valid
TEST=cros_sdk tast -verbose run --extrauseflags chrome_internal my_crbook ui.ChromeCrashNotLoggedIn.*
TEST=cros_sdk tast -verbose run --extrauseflags chrome_internal my_crbook crash.*

Change-Id: I5822f016642986d7eba5d5b14f966aa121095893
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5376100
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>