shill: wifi: trigger scan on significant RSSI drop
Keep track of the maximum connected RSSI with the current endpoint so
that significant drops can trigger a scan. This way, significant
movement will cause reevaluation of BSSes (and not just RSSI below the
scan threshold).

BUG=b:321899910
TEST=manual - deploy changes to Karis, connect to GoogleGuest, walk to
one end of the building. Observe that roaming is sometimes triggered
before the low RSSI threshold is reached.

Change-Id: Ib53565eba0509a219e6a879319fd8053998d9d13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5361938
Reviewed-by: Kevin Lund <kglund@google.com>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>