heartd: Add RunAction method and kSyndData action
Add RunAction method so that we can run a specific action immediately
for testing. This won't introduce security concern because if a client
can get HeartdControl interface, it means it can also enable the reboot
action.

This CL also adds kSyndData action, so that we can test this command in
Tast lab to make sure the seccomp policy is healthy.

BUG=b:328347355
TEST=FEATURES=test emerge-brya heartd
TEST=heartd-tool --run_action_now --action1=kSyncData

Change-Id: I540eb0a0afd3770cfa7af39a62a509277b1c4e75
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5364413
Commit-Queue: Kerker Yang <kerker@chromium.org>
Reviewed-by: Byron Lee <byronlee@chromium.org>
Tested-by: Kerker Yang <kerker@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>