diagnostics: Remove empty routine detail in mojom
The `detail` field in `RoutineStateFinished` is optional. We should put
a nullptr if a routine has no extra detail to provide.

Before this change, every routine reports a routine detail (possibliy
empty) when the routine finishes. After this change, a routine may
report a null routine detail when that routine has no extra detail to
provide.

The [MinVersion] attributes in the Mojo definition is removed because
removing members from a union type breaks the backward compatibility. In
that case, [MinVersion] is useless.

The copy of the Mojo definition in the client sides should be updated as
well because this change is not backward compatible.

BUG=b:328729519
TEST=FEATURES=test emerge-volteer diagnostics

Change-Id: Ie755e578729331007953b3819229d5567d44a946
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5364376
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Tested-by: Wei-Luan Wang <weiluanwang@google.com>