cryptohome: Clean up uss storage module
In addition to a few minor cosmetic cleanups, this also deletes some
TODO to use multiple slots of USS. There's not a lot of point to doing
them at this point because WriteFileAtomicDurable already does a "write
to temporary file, replace existing file" in a fairly synchronous way.
So using multiple slots would basically just be adding a second layer of
the same thing.

I wish we could actually make the atomic durable write even better, in
particular it's technically possible that it could leave behind a junk
file if you were to crash right at the worst time. But using USS slots
wouldn't actually help with that (in fact it would have the same
problem). So I think this is as good was we can reasonably get, barring
a switch to a much more complex approach.

BUG=b:232299885
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Id0d045040cf729142aad442d23ea988f8886ecbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5366160
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>