cryptohome: Generalize kLeLocked primary action
Recent Cryptohome changes add a lock mechanism to CryptohomeRecovery,
which is not a low entropy (LE) credential. Change kLeLocked to
kFactorLocked to universally use this for any factor.
The enum value is not referenced in the Chrome code by its string,
so it should be fine to modify the string value.

BUG=b:324076479
TEST=Cq passes.

Cq-Depend: chromium:5367067
Change-Id: I9632877a2292d06f6ce659cf02785919031346d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5367146
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: John L Chen <zuan@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>