shill: cellular: avoid crash when running in some test envs
For example, if you are trying to run shill unit tests from
/var/cache under gdb using platform2_test, the CHECK_NE would
cause a spurious segfault, presumably because the
decltype(nullptr) of its right-hand side would be reinterpreted as
const char * and then the log message would try to print it as a
string, causing a crash.

The test still fails after changing this out for ASSERT_NE, but it
fails with an understandable message now. We could further mitigate
this by changing the throw_on_failure gtest flag to false, or
hopefully this will make it more obvious that the user trying to run
tests under gdb should use --gtest_filter to skip these tests.

BUG=None
TEST=unit tests via FEATURES=test emerge (pass normally), as well as:
  $ cros_sdk -- \
      /mnt/host/source/src/platform2/common-mk/platform2_test.py \
      --board=${BOARD} /bin/bash
  -> $ /var/cache/portage/chromeos-base/shill/out/Default/shill_unittest
  which fails, but no longer crashes by segfaulting

Change-Id: I3f185f49051f1d27c086a89e02ec7c526ce018d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5367660
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>