hiberman: Log error if VolumeManager singleton can't be created
VolumeManager::new() returns an error if it can't determine the
volume group of the first partition of the root storage device.
VolumeManager is a singleton that is instantiated in a
'lazy_static' block. If the creation of the VolumeManager fails
the block currently panics and the error is printed in the panic
message, but the panic message isn't included in the syslog. Log
the error besides panicking to generate a syslog entry.

BUG=b:302612046
TEST=force error in VolumeManager::new()
  => the error is logged in the syslog

Change-Id: I5802848f6afd9a1437d1bdcaa45d2bb8dbd7c036
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5368552
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>