vm_tools: concierge: check for VM mitigations earlier
If the host machine is missing the virtualization-related security
mitigations, there is no reason to check anything else, so move the
check to the very top of the StartVm function.

Also rename the function to SafeToRunVirtualMachines() to clarify the
meaning since all VMs are untrusted now.

BUG=b:322252713
TEST=FEATURES=test emerge-$BOARD vm_host_tools

Change-Id: I657b4845daa339889d7d8187e6c6b4468a2a7754
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5368556
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>