resourced: schedqos: validate process timestamp on set_thread_state
This ensures that the thread belongs to the registered process and make
the library robust to PID reuse. There can be a rare race that the
original process is dead and a thread and a process with the same TID
and PID is created.

BUG=b:328715424
TEST=sudo cargo test -p schedqos

Change-Id: Ia8135f1667a02b4992cc5c69717b941dea88bb56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5369396
Reviewed-by: David Stevens <stevensd@chromium.org>
Tested-by: Shin Kawamura <kawasin@google.com>
Reviewed-by: Youssef Esmat <youssefesmat@chromium.org>
Commit-Queue: Shin Kawamura <kawasin@google.com>