cryptohome: Get rid of location_utils.h
The purpose of this header is basically just to provide a macro for
declaring locations when making errors. However, because of the way
symbols and macros interact this tends to interact poorly with linting
and other tools that are designed to figure out what headers you need to
include.

This isn't a super critical error: so long as location_utils.h includes
locations.h then everyone that uses it is going to work fine. But
there's no real reason not to simplify and just move the macro into
locations.h itself. There's no significant overhead imposed on code that
only wants to use the enum.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=ran location_db.py to make sure it still works properly

Change-Id: Ie3f2caf20b9b17400858ef321fda3c9a72d09f6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5370743
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>