arc-setup: Reland: Adopt mtk-socinfo to populate platform information
The modification leverages the new mtk-socinfo driver[1], providing more
detailed platform information which includes the ability to distinguish
between different SoC SKUs and retrieve the model names and removing the
need for hardcoded mappings for MT8183 (Kukui and Jacuzzi).

This update also handles the legacy platforms which broke CTS previously
due to the change in the model property of the released devices.

[1]:
https://lore.kernel.org/all/20231222080739.21706-1-william-tw.lin@mediatek.com/

BUG=b:293929346
BUG=b:324827119
TEST=Unittest: FEATURES=test emerge-brya arc-setup
     In android-sh: getprop ro.soc.model
     Tomato (Launched):
       MT8195
     Voltorb (Launched):
       MT8186
     Chinchou (Development):
       Kompanio 520 MT8186

Change-Id: I65584eaae79a2dd27d494a238e358a1553a8b3b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5382070
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Wei-Shun Chang <weishunc@chromium.org>
Commit-Queue: Wei-Shun Chang <weishunc@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>