hiberman: Skip hibernate if PinWeaver has credentials
Currently the PinWeaver tree is stored on the stateful partition.
When resuming from hibernate a DM snapshot device is created for
stateful, all writes are captured by the snapshot device and not
committed to disk, unless hibernate is aborted. In case of a
successful resume from hibernate the DM snapshot device and the
data it contains are discarded. This can result in
inconsistencies in the PinWeaver tree if the tree has changed
due to failed authentication attempts.

Eventually the PinWeaver tree will likely be moved to a dedicated
partition, for now abort hibernate if PinWeaver credentials
exist on the device.

Don't skip preparations for hibernate at the login time if
PinWeaver credentials exist. For one the user could remove the
credentials before hibernating, but more importantly it allows
to capture metrics of how often hibernate is skipped due to
the existence of PinWeaver credentials.

BUG=b:326563622
TEST=hibernate w/o PinWeaver credentials => ok
  try to hibernate w/ PinWeaver credentials => aborted

Change-Id: I597f68796fc7e3d75458a7b184a55e7dd1e0973b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5372593
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>