shill: cellular: remove unused "synchronous error"
This argument is unused by the real implementation of the D-Bus
proxy methods, but the mock will set it to return an error. This
means we could potentially have distinct error pathways depending
on whether or not we're running inside unit tests.

Remove the synchronous error and always use the asynchronous
callback flow.

BUG=b:172215298
TEST=unit tests

Change-Id: Iccc2f1f149b85be4cc3bf2ed4183a31584c9189a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5372804
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>