adbd: Update startup scripts for DbC to use typec usb role
DbC requires the usb port to be in host mode. On many devices this
can be set using the /sys/class/usb_role/CONX-role-switch file,
where X is the port number. However, on a number of machines
(eg. mithrax, taniks, felwinter, taeko), the CONX file does not have
the same number as the port.

In order to make sure we map to the correct control file, the
start-up script should first try to use
/sys/class/typec/portX/usb-role-switch/role which should point to the
correct control file.

BUG=b:321051330
BUG=b:121287022
TEST=Activate ADB over USB toggle in Android settings.
  No cables attached:
  cat /sys/class/typec/port1/usb-role-switch/role
  none

  Plug in cable to right-hand port (port1):
  cat /sys/class/typec/port1/usb-role-switch/role
  host

  Remove cable:
  cat /sys/class/typec/port1/usb-role-switch/role
  none

Change-Id: I47add33e52cf2cfcf7f6a01424e65d5d5b7b6681
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5374901
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Emilie Roberts <hadrosaur@google.com>
Tested-by: Emilie Roberts <hadrosaur@google.com>
Reviewed-by: Emilie Roberts <hadrosaur@google.com>