crash: Change Is64BitProcess to return std::optional
A future CL will remove CrashCollector::ErrorType, which is
ArcppCxxCollector::Is64BitProcess's current return type. As the return
value is only used as a success / failure boolean, and the more granular
failure type information is thrown away, it seems reasonable to
migrate the function to just returning a success/failure status instead
of adding enum values to the eventual replacement of
CrashCollector::ErrorType.

In keeping with modern coding standards, use std::optional to represent
that success / failure, and move the output parameter into the return
value.

This CL should have no impact on the behavior of the program.

BUG=b:177552411
TEST=Unit tests
TEST=cros_sdk tast -verbose run --extrauseflags chrome_internal my_crbook arc.AppCrash*

Change-Id: I186528571243b13ceedbba83bcd0ffb8bf470cfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5375765
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>