patchpanel: Remove interface name check in Firewall
The interface name will only be used in iptables, which should be
validated by both iptables and the kernel. There doesn't seem to be a
strong reason to keep an additional check in patchpanel, and it can
cause some confusion if the check doesn't match the one in the kernel.

BUG=b:330089306
TEST=FEATURES="test" emerge-$BOARD patchpanel

Change-Id: I26131766803c928be73a2f3f77b17bd6af1f82b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5369745
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>