mojo_service_manager: move MojoServiceProvider to mojo_service_manager
To register a mojo service at mojo service manager, we need to implement
the `ServiceProvider` interface. The implementation of ServiceProvider
is the same in most of the case, so we have a template implementation at
heartd and diagnostics package.

This CL moves the template implementation back to mojo_service_manager
package to eliminate the duplication. It could be reused by other
packages in the future.

BUG=none
TEST=FEATURES=test emerge-brya mojo_service_manager \
     chromeos-base/heartd diagnostics
TEST=`emerge-brya mojo_service_manager-test` and run
     `/usr/local/bin/mojo_service_manager_test_tool \
      --action=create-test-service` on DUT

Change-Id: Id7937bc89dba0e4aef1b24fab30e35477699ade3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5373957
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>