system-proxy: Remember the auth methods supported by the proxy server
When a web proxy server requires authentication credentials, it sends
the supported authentication methods along with the 407 HTTP response.
If system-proxy does not have credentials to authenticate, system-proxy
sends a request to Chrome to get credentials for that specific web
proxy server and the supported authentication methods.

With the current implementation, curl will send another unauthenticated
request to the remote server, wait for the web server to send a 407
response and then send an authenticated request using the most secure
authentication method supported by the server.

This CL updates system-proxy to remember the authentication methods
which the remote proxy sent in the first 407 HTTP response and use it
to authenticate when trying the newly acquired credentials.

BUG=b:320167232, b:327988700
TEST=tast run $DUT network.SystemProxyForArc
     system-proxy unit tests

Change-Id: Ia3526bc3edfa33be1cb6da7b682e7f262ed19976
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5377786
Commit-Queue: Andreea Costinas <acostinas@google.com>
Tested-by: Andreea Costinas <acostinas@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Andreea Costinas <acostinas@google.com>