iioservice: Add post-stop signal handling
Upstart treats a failing post-stop process as fatal, and avoids
respawning the service out of an abundance of caution. crrev/c/5247928
added new ways for post-stop to fail -- particularly, that if it takes
longer than the timeout (5 seconds, by default), we send it SIGTERM.

The post-stop job here only sleeps for 2 seconds, so it's unlikely to
hit a 5 second timeout. Still, to prevent any possible issue, add a
"exit 0" SIGTERM handler, so it's not treated as a failure if we ever do
reach this timeout

BUG=b:330056158
TEST=`pkill -SIGABRT iioservice`, `status iioservice` still respawns

Change-Id: I993aa209696c9b30275ac03d638346509bdecd94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5376912
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>