crosh: remove board_name() function
The only user of this function can be replaced with a static string that
just says the feature is not available on this device.

This resolves a clippy::lines_filter_map_ok lint that was present in the
implementation of the board_name() function.

BUG=None
TEST=cargo clippy

Change-Id: I91f5565f6c3186558147449371f4cbe0160ba91b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5378771
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>