diagnostics: Support interactive V2 routine
Add a field for interaction in the routine waiting state to indicate
that the routine requires the client to take actions to proceed. For
some interactions (called inquiries), clients need to call the
`ReplyInquiry` method to respond.

DD: go/telemetry-extension-interactive-routine.

Note that RoutineTask is missing from the mojom file because there is
currently no interaction that does not need a reply.

BUG=b:309781398
TEST=FEATURES=test emerge-volteer diagnostics

Cq-Depend: chromium:5276818
Change-Id: I522d98b44668fff99513da9d6eacaec1d5d8c2cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5171546
Tested-by: Wei-Luan Wang <weiluanwang@google.com>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>