cryptohome: Make a PrepareInput type and use it in the driver
This converts the existing AuthInput parameter used by the Prepare calls
in the driver over to using PrepareInput instead. Initially the prepare
input is just the subset of auth input that's currently used (by
fingerprint) but the plan is to add in the new subobjects needed by
recovery.

In theory we could just add those new inputs to the AuthInput instead
but I find that gets extra confusing, because it already has objects for
the recovery input (for the Authenticate operation). I think we mainly
reused AuthInput to avoid duplicating a whole new somewhat-duplicated
type but now that we actually need the prepare to take different inputs
I think making a new type is a better idea.

BUG=b:269619280
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I128aa64ce4a3a9e298cbbf7cd86b4ae505f750d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5381443
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: John Admanski <jadmanski@chromium.org>