shill: cellular: Re-enable IPv6 explicitly in L850+VZW in shill
Since commit 956ba97e8ec1f534 in R117 shill will uses the link-local
address specified by the modem to perform the IPv6 SLAAC procedure (in
the context of b/292048101). This fix would avoid the multiple source-ip
violations in the network, so there should be no need to disable IPv6
explicitly any more on VZW+L850.

Moreover, since commit e63c4238984 in R122 (in the context of
b/310710142), we will typically connect the Class 3 APN in Verizon in
the L850 using the profile ID instead of passing a full set of
connection settings to the modem. Due to this, we're really no longer
disabling IPv6 in the connection attempt if the stored profile settings
configure dual-stack IP type. This means that effectively we have
already removed the workaround from the logic, so it is an additional
reason why it can be removed from the code in shill.

BUG=b:330117048
TEST=tast -verbose run dut  cellular.Smoke
TEST=tast -verbose run dut  cellular.HostStressEnableDisable

Change-Id: Iae57eab81fbd70a08a09fd015af337ca8b232648
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5381841
Tested-by: Madhav <madhavadas@google.com>
Auto-Submit: Madhav <madhavadas@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>