vm_tools: sommelier: basic heuristics to filter out non-game windows
We should not apply containerized windows quirk if the window
is not a game window (we are limiting the support for thiss quirk
to game windows only, for now).

This CL has very simple heuristics to guess if the window is an
actual game window (telling apart from lauchers and anti-cheat
software windows). More work needs to be done.

BUG=b:329779634,b:330639704
TEST=Test quirks across multiple games with launchers and
anti-cheat windows.

Change-Id: Ic9e1d89330b5a19bd0aa3011c8062f521c959b79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5382299
Reviewed-by: Chloe Pelling <cpelling@google.com>
Tested-by: Max Lee <endlesspring@google.com>
Commit-Queue: Max Lee <endlesspring@google.com>