vm_tools: sommelier: set aspect ratio for containerized windows
Set the window aspect ratio to be same as the actual game window,
so that when user resizes, window does not end up with
transparent areas where viewport is not filling.

Note that with set_aspect_ratio, the title bar of the window is
taken account into ratio calculation in Exo. We should check
if this is intended. If yes, we should update the implementation
to get title bar size into account.

BUG=b:329779634,b:330639760
TEST=resize game windows

Change-Id: Ia05a68d754d953069c4dfd7de38cea7f2f75e338
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5379431
Reviewed-by: Chloe Pelling <cpelling@google.com>
Commit-Queue: Max Lee <endlesspring@google.com>
Tested-by: Max Lee <endlesspring@google.com>