cryptohome: Don't replace scheduled signal
Currently the auth factor update signals are scheduled on the same wall
clock timer. When we call Start() again, the existing scheduled event
will be dropped. It's fine to replace the scheduled event if the new
desired run time is closer to now (as all scheduled events are the
same), but if the new desired run time is further from now, we should
just skip the scheduling.

BUG=b:328347462
TEST=cryptohome unittests

Change-Id: Ifc22e137232ce4f9edef9bed0334a7d749792c3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5374240
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>