cryptohome: Add PrepareOutput to prepare tokens
This allows for the preparation process to supply output. I did this by
attaching the output struct to the token, as it's already an object that
gets handed back to the caller and so rather than creating a parallel
data channel we can just keep the data in one place.

I added the recovery field to the output, although none of the existing
tokens populate it. This will get used when I move the mechanics of the
GetRecoveryRequest operation into the recovery driver.

BUG=b:269619280
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Iaa3e0dd218b809d7474b0af2e4ae57cf82f9bbeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5383558
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>