shill: Add Priority class to WiFiPhy
WiFiPhy will be the owner of concurrency decision making logic, so it
should also own some basic functionality to process priority values from
API calls. This CL adds a Priority class to WiFiPhy which is a thin
wrapper on a uint32_t. This class contatins the minimum and maximum
priority values as defined in go/cros-wifi-concurrency.

BUG=b:328074924
TEST=FEATURES=tests emerge-brya shill

Change-Id: Id9a7c6982f1dee10b5c9a95352109569d1dfb5d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5386502
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>