init: compile mount_encrypted depends on USE encrypted_stateful
mount_encrypted is not needed when USE flag encrypted_stateful is not
set.
Move of mount_encrypted to init exposes it for viking, it was previously
in cryptohome that is not compiled on that board.

BUG=b:172221451
TEST=emerge-viking chromeos-init. CQ.

Fixes: commit f58688a4ba ("init: move mount_stateful")
Change-Id: I980746055acd6ef5dcf69acab5529c0fd7010838
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5384701
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Baili Deng <bailideng@google.com>
Auto-Submit: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Baili Deng <bailideng@google.com>