cryptohome: Create a cryptohome recovery auth block service
This creates a recovery "service" in cryptohome that operates similarly
to the fingerprint services, in the sense that it will be responsible
for handling Prepare operations and for generating tokens. It's not
really truly a service in the same way as the others in that it doesn't
have the same kind of state to track, all of the relevant state is
contained in the token it generates, but it seems easiest to understand
if we reuse the same pattern.

In this case we don't (yet) actually use it for Prepare, or as a
service. For now this just relocates the innards of the operation from
AuthBlockUtility to the new service. Eventually we'll move the call to
happen with Prepare but this moves things in that direction
incrementally.

BUG=b:269619280
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ic449db6fc157ee67b26adbbe8f1020ae61be2761
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5386445
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>