print_tools: Remove memory leak in MockResolver
freeaddrinfo() frees the passed-in structure, but doesn't free the
memory pointed to by ai_addr.  This causes structures allocated in the
MockResolver to leak a pointer during tests.  Fix this by allocating a
single large block of memory and pointing ai_addr into the same block
after the end of struct addrinfo.  This doesn't affect non-test code
because the real getaddrinfo allocates memory in a way that freeaddrinfo
expects.

BUG=b:330450277
TEST=unit tests on amd64-generic-asan

Change-Id: Ifffae536ac6d072a030806f6444262600bd711fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5386382
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>