shill: Sort ConcurrencyCombinations by num_channels
This is a minor efficiency optimization to speed up concurrency
decision making while keeping the code simple.

In an upcoming concurrency CL, ConcurrencyCombinations are processed to
determine the valid combination which supports the maximum number of
channels. Sort the combinations by number of channels so that we can
process them in order and stop whenever we find a valid combination.

BUG=b:328073403
TEST=FEATURES=test emerge-brya shill

Change-Id: Ib6b675d59a88d47274aa152bd44dd14dccbf820d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5398173
Reviewed-by: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>