vm_tools: sommelier: remap iconified windows upon gaining focus
Initially, unmap-map happens in a single execution, which can lead to
janky window screens (flickering when losing focus) and crashes Ash
when window is unmapped but Exo requests are processed.

By moving mapping into focus event, we mitigate the issues above.
However, this means the game will be blackscreen in overview mode
since it has entered 'iconified' state. I think this is okay, since
the developers doing this intentially wanted their game to not render
when minimized.

This is a mitigation in-place until we fully investgate why Ash
crashes in the scenario specified. We should probably safe-guard
instead in the long-run.

BUG=b:330652242
TEST=HADES and Euro Truck Simulator 2 do not have black screen and
do not crash with containerization.

Change-Id: I8be54d08b3da4854d37a17e666f0d118e7186f61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5386550
Commit-Queue: Max Lee <endlesspring@google.com>
Reviewed-by: Chloe Pelling <cpelling@google.com>
Tested-by: Max Lee <endlesspring@google.com>