shill: Split SetTetheringEnabled into two functions
Split SetTetheringEnabled into two functions EnableTethering and
DisableTethering. The underlying logic is unchanged, so EnabledTethering
should be functionally identical to SetTetheringEnabled(true) and
DisableTethering should be functionally identical to
SetTetheringEnabled(false).

EnableTethering takes uint32_t arg "priority" which is mandatory and
will be used for concurrency resolution. It is unused in this patch.

SetTetheringEnabled is now designated as deprecated, but will be left in
the code until the API callers have been updated to use the new
functions.

BUG=b:328074924
TEST=FEATURES=test emerge-brya shill
TEST=Use new API commands to start/stop tethering via dbus.

Change-Id: I87de2c578032cfe9251b5071aee7a18ab1ffb35b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5398170
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>