hwsec-host-utils: x-ver: allow multiple boards/versions
In most of the case, we would need prepare for multiple boards for a
milestone. By allowing multiple boards for --board, we would not need to
manually run multiple bash commands (or loop) when preparing x-ver data.

Although ideally we only need to prepare for the latest stable milestone
per month, it's still possible that we need to manually prepare for
multiple milestones/versions. In such case, allowing multiple
versions/milestones in --version/--milestone should be helpful.

BUG=b:331504778
TEST=manual

Change-Id: I66249f20d099efd3b3fa7477455fa61a85bcff4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5391251
Reviewed-by: John L Chen <zuan@chromium.org>
Tested-by: Ching-Kang Yen <chingkang@chromium.org>
Commit-Queue: Ching-Kang Yen <chingkang@chromium.org>