shill: vpn: Parse pppd log for authentication failure
We use the exit code of pppd to detect whether there is an
authentication failure during the connection, but this may have false
negative. In some cases, the peer connection is closed at first after
the auth failure, and then the xl2tp on the local side gets this event
and terminate the pppd. In this case, pppd will exit with 16
(EXIT_HANGUP) instead.

This patch adds the logic to store the pppd log in a temp file and after
the connection ended, read and parse this file to check if there is an
auth failure.

BUG=b:329328608
TEST=FEATURES="test" emerge-$BOARD shill
TEST=`tast run $DUT network.VPNIncorrectCreds*` on brya, it's failing
  before this patch and passing after this patch.

Change-Id: I78413368875a7b3b90b1ac2472b11278d668cf56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5388401
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>