cryptohome: Implement prepare for recovery factor
This adds a singleton instance of the cryptohome auth block service and
hooks it up to the recovery auth factor driver.

The implementation fills in the driver implementation fairly fully and
also expands the PrepareInput to take the necessary parameters. It stops
short at adding the support to the AuthSession Prepare call for actually
constructing the PrepareInput from the request. This is because adding
this conversion will basically require converting the existing
GetRecoveryRequest implementation over to using Prepare, leading to a
bunch of additional changes to align with that. This change stops just
short of that, with the plan to do that in the next commit.

This change also virtualizes the auth block service so that we can make
a mock. In general I'm not a huge fan of this approach, but doing a
whole fake recovery request in the unit tests at every layer up the
stack just doesn't scale well, it's far too much setup (even using the
TPM simulator, which not every test can use). For tests of the higher
layers it just became more practical to make the underlying
GenerateRecoveryRequest mockable.

BUG=b:269619280
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I3a36d48b96574ccfdf558682da4b48579548d5f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5399320
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>