shill: portal: define "http-only" mode
This patch adds to PortalDetector a boolean option to only send the HTTP
probe for detection a portal without sending the HTTPS probe validating
Internet connectivity and without reporting kNoConnectivity results.

This mode is useful for implementing the "http_only" mode of the Service
CheckPortal property.

BUG=b:318370676
TEST=FEATURES=test emerge-$BOARD shill. Manual: flashed shill with
crrev/c/5409955, set CheckPortal to "http-only" on GoogleGuest, verified
in /var/log/net.log and with tcpdump that a single HTTP request is sent.

Change-Id: Ib65124f61590fa031569649c324c3799608f674c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5404318
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>